+++ test/Main.hs
@@ -12,6 +12,7 @@ import Tests
   , test3, test3b
 #endif
   , test4, test4b
+  , test5
   )
 
 --------------------------------------------------------------------------------
@@ -28,6 +29,7 @@ tests =
 #endif
   , Test "test4"  test4  "<<no MyShow instance>>"
   , Test "test4b" test4b False
+  , Test "test5"  test5  'x'
   ]
 
 main :: IO ()